Fix IssueInstantMillis to be a valid SAML instant
SAML 2.0 specification (see section 1.3.3 of core specification) requires all time values to be valid xs:dateTime values in UTC format. So, the trailing 'Z' (which indicates UTC) is needed here, otherwise compliant parsers will fail to parse this IssueInstantMillis, not because the presence of fraction seconds (which 110 test is meant for), but because of the missing 'Z'.